fixed NSFW imgs posting in SFW channels, also fixed for ~r commend, if it meses up it sends back shrug. need to implement for rest of them

+++ sarina.py
@@ -36,12 +36,18 @@ async def on_message(message):
             if message.channel.name == "nsfw" or message.channel.name == "gayboys":
                 
                 returnMessage = getSubredditPictureNSFW(message.content[3:])
-                await client.send_file(message.channel, returnMessage)
-                os.remove(returnMessage)
+                if returnMessage == ":shrug:":
+                    await client.send_message(message.channel, returnMessage)
+                else: 
+                    await client.send_file(message.channel, returnMessage)
+                    os.remove(returnMessage)
             else:
                 returnMessage = getSubredditPicture(message.content[3:])
-                await client.send_file(message.channel, returnMessage)
-                os.remove(returnMessage)
+                if returnMessage == ":shrug:":
+                    await client.send_message(message.channel, returnMessage)
+                else: 
+                    await client.send_file(message.channel, returnMessage)
+                    os.remove(returnMessage)
         ###owl attack
         elif message.content.startswith('~owl'):
             try:
@@ -72,6 +78,12 @@ async def on_message(message):
             await client.send_message(message.channel, '***Captain*** Jack Sparrow')
         elif 'love you' in message.content or 'love sarina' in message.content or 'love this bot' in message.content:
             await client.send_message(message.channel, 'Love you too')
+        elif 'bad bot' in message.content or 'bad sarina' in message.content:
+            await client.send_message(message.channel, "I'm sorry D:")
+        elif 'good bot' in message.content or 'good sarina' in message.content:
+            await client.send_message(message.channel, "Thanks :) <3")
+        elif 'thanks bot' in message.content or 'thanks sarina' in message.content or 'thanks sarina!' in message.content:
+            await client.send_message(message.channel, "You're welcome :)")
 
 
         ###for explosion
@@ -106,19 +118,24 @@ def getSubredditPicture(subreddit="", nsfw=False):
                         # and redditApiResult['domain'].lower() in imageDomains:
                             if nsfw is False and redditApiResult['over_18'] is False:
                                 gettingPicture = False
-                            elif nsfw is False:
+                            elif nsfw is True:
                                 gettingPicture = False
                     maxReTries = maxReTries - 1
                 if not gettingPicture:
+                    print(nsfw)
+                    print(redditApiResult['over_18'])
+                    print(gettingPicture)
                     replyMessage = redditApiResult['url'].replace('amp;', '')
+                    imageResponse = redditApiResult['url'].replace('amp;', '')
                     print(replyMessage)
-                    fileName = replyMessage.split('/')[-1]
+                    replyMessage = replyMessage.split('/')[-1]
                     requestHeaders = {'User-agent': 'linux:Sarina:v1'}
-                    imageResponse = requests.get(replyMessage, headers=requestHeaders, stream=True)
-                    imageFile = open('imgs/SFW/' + fileName, 'wb')
+                    imageResponse = requests.get(imageResponse, headers=requestHeaders, stream=True)
+                    imageFile = open('imgs/SFW/' + replyMessage, 'wb')
                     imageFile.write(imageResponse.content)
                     imageFile.close()
-        return 'imgs/SFW/' + fileName
+                    replyMessage = 'imgs/SFW/' + replyMessage
+        return replyMessage
     except Exception as e:
         print('Error in getSubredditPicture method:')
         print(e)
@@ -147,7 +164,7 @@ def getSubredditPictureSpecific(subreddit = '', nsfw=True):
                     if 'poop' not in redditApiResult['title'].lower() \
                         and redditApiResult['score'] > 10:
                         # and redditApiResult['domain'].lower() in imageDomains:
-                            if nsfw is True and redditApiResult['over_18'] is True:
+                            if nsfw is False and redditApiResult['over_18'] is False:
                                 gettingPicture = False
                             elif nsfw is True:
                                 gettingPicture = False
@@ -199,16 +216,20 @@ def getSubredditPictureNSFW(subreddit="",nsfw=True):
                                 gettingPicture = False
                     maxReTries = maxReTries - 1
                 if not gettingPicture:
+                    print(nsfw)
+                    print(redditApiResult['over_18'])
+                    print(gettingPicture)
                     replyMessage = redditApiResult['url'].replace('amp;', '')
-                    fileName = replyMessage.split('/')[-1]
+                    imageResponse = redditApiResult['url'].replace('amp;', '')
                     print(replyMessage)
-                    print(fileName)
+                    replyMessage = replyMessage.split('/')[-1]
                     requestHeaders = {'User-agent': 'linux:Sarina:v1'}
-                    imageResponse = requests.get(replyMessage, headers=requestHeaders, stream=True)
-                    imageFile = open('imgs/NSFW/' + fileName, 'wb')
+                    imageResponse = requests.get(imageResponse, headers=requestHeaders, stream=True)
+                    imageFile = open('imgs/NSFW/' + replyMessage, 'wb')
                     imageFile.write(imageResponse.content)
                     imageFile.close()
-        return 'imgs/NSFW/' + fileName    
+                    replyMessage = 'imgs/NSFW/' + replyMessage
+        return replyMessage   
     except Exception as e:
         print('Error in getSubredditPicture method:')
         print(e)
